merging game data into katrina

See merge request !28

+++ src/game_data/Sprite.java
@@ -15,7 +15,7 @@ public abstract class Sprite {
 	
 	private Location myLocation;
 	private String myImagePath;
-	private int myVelocity;
+	private double myVelocity;
 	private CollisionHandler myCollisionHandler;
 	private Set<Characteristic> myCharacteristics;
 	
@@ -64,10 +64,10 @@ public abstract class Sprite {
 	public void setMyLocation(Location myLocation) {
 		this.myLocation = myLocation;
 	}
-	public int getMyVelocity() {
+	public double getMyVelocity() {
 		return myVelocity;
 	}
-	public void setMyVelocity(int myVelocity) {
+	public void setMyVelocity(double myVelocity) {
 		this.myVelocity = myVelocity;
 	}
 	public String getMyImagePath() {
